Protector-Heart link
The protectors are now linked to their Hearts. When a Heart is placed on
a Invoker, the protector appears and its health is displayed above the
invoker. When the heart is taken out, the protector dispawn and the
heart saves its HP.
If a protector dies, its heart turns into a "dying heart" than can be
regenerated in the radiance catalyst.
Also updated the radiance catalyst renderer.