New fluid: Liquid corruption
This new fluid replace lava in the abyss. It is very slow and spawns naturally at the bottom of the world. It is not pickable with a bucket. It is expected that this fluid gives a negative effect when touched, somthing between fire and poison, with high damage and visual overlay.