Noctilure integration
Updated the Noctilure taming and spawn. It now spawn in the abyss, can be tamed using the new item Noctilure Treat crafted from the new Blue Glow berries dropped from the blue bushes. The Noctilure can also sprint in the air, but has an endurance that can be recharged by not sprinting or landing.