Citadel update
Remove the issue with citadel appearing cut off at a 8 chunk distance by removing the ability for the structure to generate beyond this limit. This shorten the structure, but with new shorter bridges, the result is very good, almost even better with bigger space being free thanks to the beard_box.