Crystal Resonator and Infuser
New item: the crystal resonator, a compass that can change color based on its crystal type and will be able to indicate crystal temples. New block: the crystal infuser that changes color based on its crystal type and  will be able to change the crystal type of a resonator. Those block will be found in crystal temples.